# docs/swagger.yaml
basePath: /api/v1
definitions:
  helpers.ErrResponse:
    properties:
      message:
        type: string
    type: object
  models.Bill:
    properties:
      bill_owner_id:
        type: integer
      billData:
        description: has many BillData
        items:
          $ref: '#/definitions/models.BillData'
        type: array
      billOwner:
        allOf:
        - $ref: '#/definitions/models.BillOwner'
        description: belongs to a BillOwner
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      raw_image:
        type: string
      updated_at:
        type: string
    type: object
  models.BillData:
    properties:
      bill:
        $ref: '#/definitions/models.Bill'
      bill_id:
        description: belongs to a Bill
        type: string
      created_at:
        type: string
      discount:
        type: number
      id:
        type: integer
      misc:
        type: string
      service:
        type: number
      store_name:
        type: string
      sub_total:
        type: number
      tax:
        type: number
      total:
        type: number
      updated_at:
        type: string
    type: object
...
  models.BillOwner:
    properties:
      bank_account:
        type: string
      contact:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Split Bill Swagger Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Split Bill API
  version: "1.0"
paths:
  /billdatas:
    get:
      description: Get all Bill Data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BillData'
            type: array
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Get all Bill Data
      tags:
      - billdatas
    post:
      consumes:
      - application/json
      description: Create new Bill Data
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.BillData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BillData'
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Create new Bill Data
      tags:
      - billdatas
  /billdatas/{id}:
    delete:
      description: Delete Bill Data by ID
      parameters:
      - description: data
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Delete a Bill Data by ID
      tags:
      - billdatas
    get:
      description: Get Bill Data by ID
      parameters:
      - description: data
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BillData'
            type: array
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Get a Bill Data by ID
      tags:
      - billdatas
    put:
      consumes:
      - application/json
      description: Edit Bill Data
      parameters:
      - description: data
        in: path
        name: id
        required: true
        type: integer
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.BillData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BillData'
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Edit Bill Data
      tags:
      - billdatas
...
  /bills:
    get:
      description: Get all Bills from table
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Bill'
            type: array
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Get all bills
      tags:
      - bills
    post:
      consumes:
      - application/json
      description: Create a new bill
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Bill'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Bill'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Create a new bill
      tags:
      - bills
  /bills/{id}:
    delete:
      description: Delete bill by ID
      parameters:
      - description: data
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Delete a bill by ID
      tags:
      - bills
    get:
      description: Get bill by ID
      parameters:
      - description: data
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Bill'
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Get a bill by ID
      tags:
      - bills
    put:
      description: Update bill by ID
      parameters:
      - description: data
        in: path
        name: id
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Bill'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Bill'
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Update a bill by ID
      tags:
      - bills
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"